Create question1-m1.c
just used the struct to make the code take high memory to store the variables in different variables and releasignthem on the basis of the program structure. 

deposit(), withdraw(): These functions manage transactions but were not utilized in the concise version.
calculateInterest(): Computes and adds interest to each account and prints the total interest.
printAccountDetails(): Prints details of a single account.
printAllAccounts(): Calls printAccountDetails() for each account, leading to redundant function calls.